requirements.txt
APScheduler==3.6.3
asyncio==3.4.3
click==7.1.1
daemonize==2.5.0
dnspython==1.16.0
docutils==0.16
dominate==2.5.1
eventlet==0.25.1
feather-format==0.4.1
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-MQTT==1.1.0
Flask-SocketIO==4.2.1
future==0.18.2
gevent==1.4.0
greenlet==0.4.15
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.1
jsonify==0.5
lockfile==0.12.2
MarkupSafe==1.1.1
monotonic==1.5
numpy==1.18.3
paho-mqtt==1.5.0
pandas==1.0.3
pip-tools==5.1.0
pyarrow==0.17.0
PyMySQL==0.9.3
python-daemon==2.2.4
python-dateutil==2.8.1
python-engineio==3.12.1
python-socketio==4.5.0
pytz==2019.3
PyYAML==5.3.1
schedule==0.6.0
serial==0.0.97
six==1.14.0
tzlocal==2.0.0
visitor==0.1.3
Werkzeug==1.0.0
